-- fixed minor issues for the newly added NSV media format
git-svn-id: https://svn.rtmpd.com/crtmpserver/trunk@372 01689507-d990-df11-9d7d-0017a411d416

+++ sources/thelib/include/mediaformats/nsv/nsvdocument.h
@@ -34,7 +34,7 @@ private:
 	Variant _metadata;
 	vector<MediaFrame> binaryHeaders;
 
-	uint8_t _framerate;
+	double _framerate;
 	uint32_t aux_plus_vidlen;
 	uint8_t num_aux;
 	uint16_t audio_len;
